+++ config/medialazyload.php
@@ -1,27 +1,27 @@
 <?php
 /*
- * This file is part of ImageLazyLoad.
+ * This file is part of MediaLazyLoad.
  *
  * (c) 2024 Debjyoti Kar <debjyotikar001@gmail.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
- 
+
 return [
   /*
   |--------------------------------------------------------------------------
-  | Enable Image Lazy Loading
+  | Enable Media Lazy Loading
   |--------------------------------------------------------------------------
   |
   | This option allows you to enable or disable lazy loading functionality
-  | for images in your application. If set to true, lazy loading will be
-  | active. If set to false, images will load normally without lazy loading.
+  | for media in your application. If set to true, lazy loading will be
+  | active. If set to false, media will load normally without lazy loading.
   |
   | Default: true
   |
   */
-  'enabled' => env('IMGLAZYLOAD_ENABLED', true),
+  'enabled' => env('MEDLAZYLOAD_ENABLED', true),
 
   /*
   |--------------------------------------------------------------------------
@@ -36,9 +36,9 @@ return [
   | Default: false
   |
   */
-  'jquery' => env('IMGLAZYLOAD_JQUERY', false),
-  'jqueryUrl' => env('IMGLAZYLOAD_JQUERY_URL', 'https://code.jquery.com/jquery-3.7.1.min.js'),
-  
+  'jquery' => env('MEDLAZYLOAD_JQUERY', false),
+  'jqueryUrl' => env('MEDLAZYLOAD_JQUERY_URL', 'https://code.jquery.com/jquery-3.7.1.min.js'),
+
   /*
   |--------------------------------------------------------------------------
   | Allowed Environments
@@ -52,7 +52,7 @@ return [
   | Default: local,production,staging
   |
   */
-  'allowed_envs' => env('IMGLAZYLOAD_ALLOWED_ENVS', 'local,production,staging'),
+  'allowed_envs' => env('MEDLAZYLOAD_ALLOWED_ENVS', 'local,production,staging'),
 
   /*
   |--------------------------------------------------------------------------